+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add process lowering pass. ([#93](https://github.com/fabianschuiki/llhd/issues/93))
 - Add desequentialization pass. ([#103](https://github.com/fabianschuiki/llhd/issues/103))
 - Add `reg` gating conditions. ([#105](https://github.com/fabianschuiki/llhd/issues/105))
+- llhd-opt: Add `-l` option to lower from behavioural to structural LLHD.
 
 ### Changed
 - Use dense vector table instead of hash map for blocks, instructions, values, and external unit data.